+++ HISTORY.rst
@@ -10,6 +10,9 @@ New Features
 - Add a new attribute to ``EHydro`` class called ``survey_grid``.
   It's a ``geopandas.GeoDataFrame`` that includes the survey grid
   of the eHydro dataset which is a 35-km hexagonal grid.
+- Add ``NFHL`` class within ``nfhl.py`` module to access FEMA's
+  National Flood Hazard Layer (NFHL) using six different ArcGISRESTFul
+  services.
 
 Internal Changes
 ~~~~~~~~~~~~~~~~